FG-14: Create initial CI/CD pipeline
There does not currently exist a pipeline which will be run
upon modification of the main branch. We need to have one such that
the main branch is protected from broken code.

This commit creates a yaml file which is found under the
.github/workflows directory, and is run on every push to main,
serviced by github actions. An initial testing script is also
created, found under the ci directory, but the contents of this
will be modified when there are actual projects and code to be tested.

Testing: None
JIRA: https://fun-software.atlassian.net/jira/software/projects/FG/boards/1?selectedIssue=FG-14