FG-20: Create CI for server
As of right now, it's hard to test CI changes without first pushing to
main, because there is no action trigger on the ci branch. Also, the server
goes untested. This is problematic, because buggy code can be pushed to main
and theoretically deployed.

This commit fixes this by adding CI to the same execution pipeline as
main. Thus, anything that would be executed in main, is also executed in
CI. Also, the server directory structure is well defined, and as such,
CI does not need to make any changes to ensure that the server is tested
before integration.

Testing: Pushed to CI, and verified test success before integration.
JIRA: https://fun-software.atlassian.net/browse/FG-20