chore: Create flatbuffers workspace
We need a way to serialize and deserialize data in our
client<->server communications. Flatbuffers allow for a very
efficient solution to this.

This change adds the flatbuffers workspace to the repo,
and adds some basic example entities to reference. It also
adds logic to the client to auto-gen the TS-defs for the
flatbuffers.

Jira FG-40