chore(server): Create initial UDP server
We need to have a foundation for the server to work from.
The foundation needs to be able to initialize a main thread
that will spawn worker threads which create a websocket for
each connection. The server also needs to be able to
receive state updates from the client, and publish state
updates to the clients on some specified tick rate.

This change introduces a package which uses tokio as an
asynch runtime. The server accepts a connection on some
ip:port. It creates an entity map from user id to their
player's state. Then, whenever an update comes in from the player,
the server derefs their id from the socket num, and updates
a "global"-ish variable containing all player states. Following this,
on the next tick, the server serializes the player state values
(for all players), and broadcasts those changes for all users. For now,
this serialization is simply json. This needs to be changed to flatbuffer.

JIRA: FG-37