chore(server): Compile server flatbuffers
The server did not have our existing flatbuffers compiled for use, nor
did we have an easy way to sync our flatbuffers to the server when they
are altered.

This change includes an initial compilation of the flatbuffers for the
server, and also creates a script 'sync_flatbuffers.sh' that can be
executed to re-compile them easily.

Jira FG-44