feat(client): added supabase auth and session storage
We needed some auth and session storage for the client to differentiate who is connected. This will
be important for all functionality, but for the moment, it's needed for testing server-client
messaging.

We solved this need by utilizing Supabase's built in auth and session context. This is a first
draft and I still expect to abstract out the different state modals.

JIRA:FG-54