feat: client<>server architecture init
In order to sync state from the browser to the server we need
bi-directional communication. In addition, we have 3 different
modes of communication to consider -- namely chat, game-state,
and standard API calls. Each use case has different requirements
and restrictions.

To solve this problem, we have an architecture that involves
three different protocols for communication. We start with a
basic HTTP API that the client can send requests to; these requests
will lead to an intent to join a game, upon which a response is
provided by the server that contains two new socket addresses for
the client to connect to. One is a websocket address, used for chat.
The other is a webRTC peer connection's socket address, which uses
UDP-like communication for **fast** game-state updates. Refer to the
linked diagram below for an illustration of the architecture:

diagram: https://docs.google.com/drawings/d/14-zXa5DRlzQ4YzlZObNpamCatTq0tzTERYv5-aar_ho/edit?
Co-authored-by: donovanphenry <donovanphenry@gmail.com>
Co-authored-by: Christopher Sterza <ChristopherSterza@gmail.com>
Jira FG-57
Jira FG-58
Jira FG-50
Jira FG-51
Jira FG-55
Jira FG-53
Jira FG-47
Jira FG-45