feat: Updated GameState, Player flatbuffers
The server will need an object with which to maintain the state of the
game. Currently, the GameState object acts as little more than a wrapper
of the Game object, which only houses metadata.

In this commit, I've updated the GameState object to house Player
information, updated the Players flatbuffer to house some more useful
information, done away with the PlayerRoles, and updated the server code
to be compatible with all of these changes.

JIRA: FG-75