feat(client): start game from lobby (#19)
Currently there is no way to actually enter a game.

This change adds the ability to start a game from the lobby. This will currently
bring you to a placeholder scene with minimal controls. This change also made minor semantic changes to indicate that the global wrapping 'Scene' element is actually a simple Canvas elem that wraps the 'scene' proper.

resolves FG-76